tests/smart_gtest/requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.13
langchain-core>=0.1.0
langgraph>=0.0.55
langgraph-checkpoint-postgres>=2.0.0
langgraph-checkpoint-sqlite>=2.0.0
openai>=1.0.0
sqlalchemy>=2.0.0
psycopg>=3.2.0
rich>=13.0.0
python-dotenv>=1.0.0 